Fix syntax error in write_to_nrrd() declaration; refactor DoseComparison class into a separate file
